+++ tests/VersionCollectionTest.php
@@ -7,25 +7,25 @@ namespace Version\Tests;
 use PHPUnit\Framework\TestCase;
 use Version\Exception\InvalidVersionString;
 use Version\Tests\TestAsset\VersionIsIdentical;
-use Version\Tests\TestAsset\VersionsCollectionIsIdentical;
-use Version\VersionsCollection;
+use Version\Tests\TestAsset\VersionCollectionIsIdentical;
+use Version\VersionCollection;
 use Version\Version;
 use Version\Comparison\Constraint\OperatorConstraint;
 
-class VersionsCollectionTest extends TestCase
+class VersionCollectionTest extends TestCase
 {
     /**
      * @test
      */
     public function it_can_be_created_via_constructor(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
         );
 
-        $this->assertThat($versions, new VersionsCollectionIsIdentical([
+        $this->assertThat($versions, new VersionCollectionIsIdentical([
             [1, 0, 0, null, null],
             [1, 1, 0, null, null],
             [2, 3, 3, null, null],
@@ -37,12 +37,12 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_can_be_created_from_version_strings(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
         );
 
-        $this->assertThat($versions, new VersionsCollectionIsIdentical([
+        $this->assertThat($versions, new VersionCollectionIsIdentical([
             [1, 1, 0, null, null],
             [2, 3, 3, null, null],
         ]));
@@ -54,7 +54,7 @@ class VersionsCollectionTest extends TestCase
     public function it_forwards_invalid_version_string_exception(): void
     {
         try {
-            new VersionsCollection(
+            new VersionCollection(
                 Version::fromString('1.1.0'),
                 Version::fromString('invalid')
             );
@@ -70,7 +70,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_is_countable(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
@@ -84,7 +84,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_provides_is_empty_check(): void
     {
-        $versions = new VersionsCollection();
+        $versions = new VersionCollection();
 
         $this->assertTrue($versions->isEmpty());
     }
@@ -94,7 +94,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_gets_first_version(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
@@ -111,7 +111,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_gets_last_version(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
@@ -128,7 +128,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_doesnt_return_first_last_versions_if_empty(): void
     {
-        $versions = new VersionsCollection();
+        $versions = new VersionCollection();
 
         $this->assertNull($versions->first());
         $this->assertNull($versions->last());
@@ -139,7 +139,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_is_iterable(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.3.3')
@@ -155,7 +155,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_is_sorted_in_ascending_order_by_default(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('2.3.3'),
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
@@ -181,7 +181,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_can_be_sorted_in_descending_order(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('2.3.3'),
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0')
@@ -205,7 +205,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_filters_versions_that_match_constraint(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.0.1'),
             Version::fromString('2.0.0'),
@@ -214,7 +214,7 @@ class VersionsCollectionTest extends TestCase
 
         $versions2 = $versions->matching(OperatorConstraint::fromString('>=2.0.0'));
 
-        $this->assertThat($versions2, new VersionsCollectionIsIdentical([
+        $this->assertThat($versions2, new VersionCollectionIsIdentical([
             [2, 0, 0, null, null],
             [2, 0, 1, null, null],
         ]));
@@ -225,7 +225,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_gets_first_last_items_after_filtering(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.0.1'),
             Version::fromString('2.0.0'),
@@ -243,7 +243,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_can_become_empty_after_filtering_out_all_versions(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.0.1'),
             Version::fromString('1.1.0')
@@ -259,7 +259,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_finds_major_releases(): void
     {
-        $releases = new VersionsCollection(
+        $releases = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.0.0'),
@@ -270,7 +270,7 @@ class VersionsCollectionTest extends TestCase
 
         $majorReleases = $releases->majorReleases();
 
-        $this->assertThat($majorReleases, new VersionsCollectionIsIdentical([
+        $this->assertThat($majorReleases, new VersionCollectionIsIdentical([
             [1, 0, 0, null, null],
             [2, 0, 0, null, null],
             [3, 0, 0, null, null],
@@ -282,7 +282,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_finds_minor_releases(): void
     {
-        $releases = new VersionsCollection(
+        $releases = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.0.0'),
@@ -292,7 +292,7 @@ class VersionsCollectionTest extends TestCase
 
         $minorReleases = $releases->minorReleases();
 
-        $this->assertThat($minorReleases, new VersionsCollectionIsIdentical([
+        $this->assertThat($minorReleases, new VersionCollectionIsIdentical([
             [1, 1, 0, null, null],
             [2, 1, 0, null, null],
         ]));
@@ -303,7 +303,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_finds_patch_releases(): void
     {
-        $releases = new VersionsCollection(
+        $releases = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.0.1'),
             Version::fromString('2.0.0'),
@@ -312,7 +312,7 @@ class VersionsCollectionTest extends TestCase
 
         $patchReleases = $releases->patchReleases();
 
-        $this->assertThat($patchReleases, new VersionsCollectionIsIdentical([
+        $this->assertThat($patchReleases, new VersionCollectionIsIdentical([
             [1, 0, 1, null, null],
             [2, 0, 1, null, null],
         ]));
@@ -323,7 +323,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_finds_latest_major_release(): void
     {
-        $releases = new VersionsCollection(
+        $releases = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.1.0'),
             Version::fromString('2.0.0'),
@@ -345,7 +345,7 @@ class VersionsCollectionTest extends TestCase
      */
     public function it_can_be_converted_to_an_array(): void
     {
-        $versions = new VersionsCollection(
+        $versions = new VersionCollection(
             Version::fromString('1.0.0'),
             Version::fromString('1.0.1'),
             Version::fromString('1.1.0')