+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - PHPUnit 8 is now the minimum required version
+- Rename `VersionsCollection` to `VersionCollection`
 - Rename `Version\Comparator` namespace to `Version\Comparison`
 - Move `Version\Constraint` namespace to `Version\Comparison\Constraint`
 - Rename `Version\Comparison\ComparatorInterface` to `Version\Comparison\Comparator`